+++ crates/compiler/src/codegen/generate_cargo_toml.rs
@@ -8,6 +8,9 @@ use legion::{Query, systems::CommandBuffer, world::SubWorld};
 use crate::{BuildContext, FeatureFlags, codegen::File, lowering::ProcBlock, parse};
 
 const REPO: &'static str = "https://github.com/hotg-ai/rune";
+/// The version of core crates that we want to target.
+const CORE_VERSION: &'static str = hotg_rune_core::VERSION;
+const PROC_BLOCK_VERSION: &'static str = hotg_rune_proc_blocks::VERSION;
 
 /// Generate a `Cargo.toml` file which includes all the relevant dependencies
 /// for this crate.
@@ -19,6 +22,23 @@ pub(crate) fn run(
     #[resource] features: &FeatureFlags,
     query: &mut Query<&ProcBlock>,
 ) {
+    if CORE_VERSION.contains("-dev") && features.rune_repo_dir.is_none() {
+        let msg = "
+            It looks like you are using a development version of \"rune\", but
+            haven't specified a \"rune_repo_dir\". Internal crates are resolved
+            using the \"$CORE_VERSION\" version from crates.io and builtin
+            proc-blocks are found using the \"v$CORE_VERSION\" tag from the Rune
+            repo, so there is a good chance you'll get compile errors about
+            unresolved dependencies. Specify the \"rune_repo_dir\" to resolve
+            this.
+        ";
+        log::warn!(
+            "{}",
+            msg.replace("\n", " ")
+                .replace("$CORE_VERSION", CORE_VERSION)
+        );
+    }
+
     let proc_blocks = query.iter(world);
     let mut manifest =
         generate_manifest(proc_blocks, &ctx.name, &ctx.current_directory);
@@ -100,18 +120,20 @@ where
     deps.insert(String::from("lazy_static"), lazy_static);
 
     // We'll always use the following HOTG dependencies.
-    let hotg_dependencies = &[
-        "hotg-rune-core",
-        "hotg-rune-proc-blocks",
-        "hotg-runicos-base-wasm",
-    ];
-
-    for name in hotg_dependencies {
-        deps.insert(
-            name.to_string(),
-            Dependency::Detailed(git_tagged_dependency(REPO, "nightly")),
-        );
-    }
+    deps.insert(
+        "hotg-rune-core".to_string(),
+        Dependency::Simple(hotg_rune_core::VERSION.to_string()),
+    );
+    deps.insert(
+        "hotg-rune-proc-blocks".to_string(),
+        Dependency::Simple(hotg_rune_proc_blocks::VERSION.to_string()),
+    );
+    // FIXME: We should probably use the actual version number instead of
+    // assuming it'll be in sync with core.
+    deps.insert(
+        "hotg-runicos-base-wasm".to_string(),
+        Dependency::Simple(hotg_rune_core::VERSION.to_string()),
+    );
 
     for proc_block in proc_blocks {
         let dep = proc_block_dependency(&proc_block.path, current_dir);
@@ -127,7 +149,8 @@ fn proc_block_dependency(
     current_dir: &Path,
 ) -> DependencyDetail {
     if is_builtin(path) {
-        return git_tagged_dependency(REPO, "nightly");
+        let tag = format!("v{}", PROC_BLOCK_VERSION);
+        return git_tagged_dependency(REPO, &tag);
     } else if path.base.starts_with('.') {
         return local_proc_block(path, current_dir);
     }
@@ -245,40 +268,7 @@ fn path_dependency(path: impl AsRef<Path>) -> Dependency {
 }
 
 fn patch_hotg_dependencies(hotg_repo_dir: &Path, manifest: &mut Manifest) {
-    let known_paths = &[
-        ("hotg-rune-core", "crates/rune-core"),
-        ("hotg-rune-proc-blocks", "proc-blocks/proc-blocks"),
-        ("hotg-runicos-base-wasm", "images/runicos-base/wasm"),
-    ];
     let mut overrides = BTreeMap::new();
-
-    let proc_blocks_dir = hotg_repo_dir.join("proc-blocks");
-
-    for (name, dep) in &manifest.dependencies {
-        let uses_hotg_github =
-            dep.git().map(|repo| repo == REPO).unwrap_or(false);
-
-        if !name.starts_with("hotg-") && !uses_hotg_github {
-            continue;
-        }
-
-        // We're pretty sure this is a hotg crate, now we need to figure out
-        // which local crate to redirect to. First we'll check a list of known
-        // crates, otherwise we'll assume it is a proc block.
-
-        let path = known_paths
-            .iter()
-            .find_map(|(n, p)| {
-                if name == *n {
-                    Some(hotg_repo_dir.join(p))
-                } else {
-                    None
-                }
-            })
-            .unwrap_or_else(|| proc_blocks_dir.join(name));
-
-        overrides.insert(name.clone(), path_dependency(path));
-    }
 
     overrides.insert(
         "hotg-rune-core".to_string(),
@@ -286,7 +276,16 @@ fn patch_hotg_dependencies(hotg_repo_dir: &Path, manifest: &mut Manifest) {
     );
     overrides.insert(
         "hotg-rune-proc-blocks".to_string(),
-        path_dependency(hotg_repo_dir.join("proc-blocks").join("proc-blocks")),
+        path_dependency(hotg_repo_dir.join("crates").join("proc-blocks")),
+    );
+    overrides.insert(
+        "hotg-runicos-base-wasm".to_string(),
+        path_dependency(
+            hotg_repo_dir
+                .join("images")
+                .join("runicos-base")
+                .join("wasm"),
+        ),
     );
 
     manifest
@@ -316,27 +315,18 @@ mod tests {
         assert!(got.contains_key("hotg-rune-proc-blocks"));
         assert!(got.contains_key("hotg-runicos-base-wasm"));
 
-        // All hotg dependencies should use the "nightly" tag from GitHub
-        for (_, dep) in got.iter().filter(|(key, _)| key.starts_with("hotg-")) {
-            let DependencyDetail {
-                version,
-                path,
-                git,
-                branch,
-                tag,
-                rev,
-                features,
-                ..
-            } = dep.detail().unwrap();
-
-            assert_eq!(git.as_deref(), Some(REPO));
-            assert_eq!(tag.as_deref(), Some("nightly"));
-            assert!(version.is_none());
-            assert!(rev.is_none());
-            assert!(features.is_empty());
-            assert!(branch.is_none());
-            assert!(path.is_none());
-        }
+        assert_eq!(
+            got["hotg-rune-core"].clone(),
+            Dependency::Simple(CORE_VERSION.to_string())
+        );
+        assert_eq!(
+            got["hotg-rune-proc-blocks"].clone(),
+            Dependency::Simple(PROC_BLOCK_VERSION.to_string())
+        );
+        assert_eq!(
+            got["hotg-runicos-base-wasm"].clone(),
+            Dependency::Simple(CORE_VERSION.to_string())
+        );
     }
 
     #[test]
@@ -344,7 +334,7 @@ mod tests {
         let path = "hotg-ai/rune#proc_blocks/modulo".parse().unwrap();
         let should_be = DependencyDetail {
             git: Some(REPO.to_string()),
-            tag: Some("nightly".to_string()),
+            tag: Some(format!("v{}", CORE_VERSION)),
             ..empty_dependency_detail()
         };