+++ crates/proc-blocks/tests/smoke_test.rs
@@ -5,7 +5,7 @@ use hotg_rune_proc_blocks::{
     Dimension, Dimensions, ProcBlock, ProcBlockDescriptor, TensorDescriptor,
     Transform, TransformDescriptor,
 };
-use hotg_rune_core::{Tensor, element_type::ElementType};
+use hotg_rune_core::{Tensor, ElementType};
 
 /// A dummy proc block.
 ///